Update README.md
Description:

The second project will implement a Microprogrammed Instruction Set Processor.

This builds on the VHDL model from the first project.

The following modifications are required:

Increase the number of registers in the register-file from 32 to 33.

This requires an additional select bit for the two multiplexers (Bus A and Bus

B) and the destination decoder. These are separate signals (TD, TA, TB) from

the Control Memory. See Figure 1. The size of the registers in the register-file